[PLAY-2416] Refactor Playbook SCSS to remove expensive rules and selectors (#5029)
With Safari 18.6, we noticed substantial slowdown in initial page load
in Nitro which rendered many of its pages unusable. The issue has been
narrowed down to a Webkit update that is part of Safari internals. As we
wait on a fix from there, this PR is our way of mitigating the effects
to improve Nitro performance.

Part of the issue with the slowdown is related to Safari no longer being
able to process 'complex' css on page load causing the delay. The
Playbook PR here refactors 90% of the css files exported from Playbook
to remove expensive css selectors such as class*=“target, etc. Below is
a list of which kits have only refactored scss files vs the ones that
have had changes made to the classnames used within kits.

## Kit changes tracked

**Only Scss Changes** = Only scss files were refactored to replace
expensive seelectos, No Change was made to logic in the kit at all. Kits
with ✅ icon did not need a refactor, they were already flattened.

**Classname Generation Changed** = Logic for generating classnames
changed. the concatenated logic was creating too much unnecessary
complexity, impossible to targwt every possible version of the classname
without using 'bad' selectors. These files now have multiple separated
classnames as opposed to a single concatenated classname. **NOTE** if
places in Nitro/Tempo are targetting classnames to overwirte CSS, attach
JS or within spec tests, those may break for these kits.

**Kits Not Touched** = These retain old code that needs to be refactored
in the future



| # | **Only SCSS Changes** | **Classname Generation Changed** | **Kits
Not Touched** |

|----|------------------------|----------------------------------|-----------------------|
| 1 | Title Detail | Flex | From Group |
| 2 | Advanced Table | Layout | Dropdown |
| 3 | Avatar | Button | Icon Stat Value |
| 4 | Badge | Title | List |
| 5 | Background | Section Separator | MultipleUsersStacked |
| 6 | Bar Graph ✅ | Form Pill | Nav |
| 7 | Body | Online Status | ProgressStep |
| 8 | BreadCrumb | timestamp | Table |
| 9 | Button Toolbar | tooltip | timeline |
| 10 | Caption | Card | typeahead |
| 11 | Circle Chart ✅ | | |
| 12 | Checkbox | | |
| 13 | Circle Icon Button | | |
| 14 | Collapsible | | |
| 15 | Contact ✅ | | |
| 16 | Currency | | |
| 17 | Dashboard Value | | |
| 18 | Date | | |
| 19 | Date Picker | | |
| 20 | Date Range Inline | | |
| 21 | Date Range Stacked | | |
| 22 | Date Stacked | | |
| 23 | Date Time | | |
| 24 | Date Time Stacked | | |
| 25 | Date Year Stacked | | |
| 26 | Detail | | |
| 27 | Dialog | | |
| 28 | Distribution Bar | | |
| 29 | Draggable | | |
| 30 | Empty State | | |
| 31 | File Upload | | |
| 32 | Filter | | |
| 33 | Title Count | | |
| 34 | Fixed Confirmation Toast | | |
| 35 | Text Input | | |
| 36 | Select | | |
| 37 | Passphrase | | |
| 38 | Phone Number Input ✅ | | |
| 39 | Selectable Card | | |
| 40 | Selectable Card Icon | | |
| 41 | Icon ✅ | | |
| 42 | Image | | |
| 43 | Form ✅ | | |
| 44 | Gauge | | |
| 45 | Hashtag | | |
| 46 | Highlight | | |
| 47 | Home Address Street ✅ | | |
| 48 | Icon Button | | |
| 49 | Icon Circle | | |
| 50 | Icon Value | | |
| 51 | Label Pill | | |
| 52 | Pill | | |
| 53 | Label Value | | |
| 54 | Legend | | |
| 55 | Lightbox | | |
| 56 | Line Graph ✅ | | |
| 57 | Link | | |
| 58 | Loading Inline | | |
| 59 | Map | | |
| 60 | Message | | |
| 61 | MultiLevelSelect | | |
| 62 | MultipleUsers | | |
| 63 | Overlay | | |
| 64 | Pagination ✅ | | |
| 65 | Person | | |
| 66 | Person Contact ✅ | | |
| 67 | Popover ✅ | | |
| 68 | Progress Pill | | |
| 69 | Progress Simple | | |
| 70 | Radio | | |
| 71 | Rich Text Editor | | |
| 72 | Selectable Icon | | |
| 73 | Selectable List | | |
| 74 | Skeleton ✅ | | |
| 75 | Source | | |
| 76 | Star Rating | | |
| 77 | Stat Change | | |
| 78 | Stat Value | | |
| 79 | Textarea | | |
| 80 | Time | | |
| 81 | Time Stacked | | |
| 82 | Weekday Stacked | | |
| 83 | User Badge | | |
| 84 | User | | |
| 85 | Toggle | | |
| 86 | Time Range Inline | | |
| 87 | Copy Button ✅ | | |

---------

Co-authored-by: gary.kang@powerhrg.com <gary.kang@powerhrg.com>