[PLAY-2466] Dialog Kit: Fixes for global props and classname not working - React (#5124)
**What does this PR do?** 
[Runway story](https://runway.powerhrg.com/backlog_items/PLAY-2466)

- Fixes for global props and classname not working as expected for react
it
- global props and classnames were being applied to the wrong div.

The issue here was that the globalProps + the className prop were being
applied to the 'dialog _wrapper' which does not actually wrap the modal.
I moved the globalProps and className to the actual modal as can be
seen[ here](https://github.com/powerhome/playbook/pull/5124/files). I
have tested that this works.

**Heres the caveat**: the way the scss for the Dialog is set up, the
margin global prop classname is conflicting with other scss if we use
the marginTop global prop BUT, this scss issue is fixed on the [safari
fix PR ](https://github.com/powerhome/playbook/pull/5029). I tested this
exact change on the safari fix PR locally and was not getting that scss
issue!

So for this one, we need to wait for the safari PR to be merged early
next week, then merge master to this PR and test before shipping.

**Screenshots:** Screenshots to visualize your addition/change


**How to test?** Steps to confirm the desired behavior:
1. Go to '...'
2. Click on '....'
3. Scroll down to '....'
4. See addition/change


#### Checklist:
- [ ] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [ ] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.
- [ ] **TESTS** I have added test coverage to my code.
- [ ] **SEMVER** I have added a `minor`, `major`, or `patch` label for
release.
- [ ] **RC** I have added an `inactive RC` label if not an active RC.