Migrate to PAC-v2 clusters (#5153)
**What does this PR do?** 

https://runway.powerhrg.com/backlog_items/WTWR-1740

This PR migrates playbook from PAC-v1 to PAC-v2 clusters. This new
generation of clusters updates versions of PAC software stack, and is
more reliable. It also streamlines image builds and speeds up cluster
management.

I have updated all references to pac-v1 clusters to their matching
PAC-v2 equivalents. The namespaces are already available at PAC-v2, see
https://github.com/powerhome/pac-namespaces/pull/791, already merged.

There is a review env up of this PR on PAC-v2:
https://pr5153.playbook.wc.beta.gm.powerapp.cloud

**About the database:**

The database is still unused, I check that:
- There is no mysql database named after playbook
- There are no connections established from the site pods to the
database
- The site seems to function normally with the database deleted on this
review env

So we will perform no data migration of it. We are not deleting the
database because there is interest in the team to use it in the soon
future.

The database deployment used the old storage class names, that do not
exist on PAC-v2. This PR already updates that to "block", according our
newest RFC.

**What happens after the merge?**

1. This PR is merged
2. Milano will deploy the staging and production envs of playbook
website
3. We check everything is OK
4. Watchtower updates the DNS records with the PAC-v2 address,
effectively moving the website access to PAC-v2. Will need to wait for
DNS TTL to expire (at most 5 mins)

Between 1 and 4, any other merges that happen will be deployed only to
PAC-v2, but will not be visible on playbook
website until step 4 is finished.

Two DNS records should be updated for migration
- playbook.powerapp.cloud
- staging.playbook.powerapp.cloud


**Screenshots:** Screenshots to visualize your addition/change

_(There is no change to code behavior)_

**How to test?** Steps to confirm the desired behavior:

_(There is no change to code behavior)_

#### Checklist:
- [ ] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [x] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.
- [ ] **TESTS** I have added test coverage to my code.
- [x] **SEMVER** I have added a `minor`, `major`, or `patch` label for
release.
- [x] **RC** I have added an `inactive RC` label if not an active RC.