[PLAY-2420] Advanced Table: First Column Border not rendering in Safari and Chrome (#5170)
**What does this PR do?** A clear and concise description with your
runway ticket url.
[Runway Story](https://runway.powerhrg.com/backlog_items/PLAY-2420)

This PR refines the SCSS rules we put in place for when the Advanced
Table kit is put within the Card kit.

Currently, the right borders of columns within a table do not appear
when it is in a card. We are adjusting the SCSS so that as long as the
table column is not the last child column in the table, the borders will
appear as intended when the vertical border prop is set.

DO NOT MERGE BEFORE REVERTING DOC EXAMPLES

[Alpha](https://github.com/powerhome/nitro-web/pull/51865)

**Screenshots:** Screenshots to visualize your addition/change
<img width="974" height="323" alt="Screenshot 2025-09-12 at 5 33 22 PM"
src="https://github.com/user-attachments/assets/5d0b9af3-8ebd-4237-8687-e321cf61a390"
/>


**How to test?** Steps to confirm the desired behavior:
1. Go to '[adjusted doc
example](https://pr5170.playbook.beta.px.powerapp.cloud/kit_category/advanced_table?type=react#multi-header-columns-with-vertical-borders)'


#### Checklist:
- [ ] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [ ] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.
- [ ] **TESTS** I have added test coverage to my code.
- [ ] **SEMVER** I have added a `minor`, `major`, or `patch` label for
release.
- [ ] **RC** I have added an `inactive RC` label if not an active RC.