[PLAY-2542] Phone Number input kit: Fix “format as you type” logic (first digit not registering) (#5365)
**What does this PR do?** A clear and concise description with your
runway ticket url.
[Runway Story](https://runway.powerhrg.com/backlog_items/PLAY-2542)

This PR addresses an issue in the Phone Number Input kit where, if the
'formatAsYouType' prop was enabled, the input field would not register
the first keypress the user made. This occurred due to a conflict
between the event listener in UseEffect along with the 'onChange'
handler.

To address this issue, I adjusted the formatAsYouType logic to better
handle and update the input state.

**Screenshots:** Screenshots to visualize your addition/change


**How to test?** Steps to confirm the desired behavior:
1. Go to '...'
2. Click on '....'
3. Scroll down to '....'
4. See addition/change


#### Checklist:
- [x] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [x] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.
- [x] **TESTS** I have added test coverage to my code.
- [x] **SEMVER** I have added a `minor`, `major`, or `patch` label for
release.
- [ ] **RC** I have added an `inactive RC` label if not an active RC.